Rename Oracle check
