+++ pkg/collector/corechecks/oracle/activity_queries.go
