+++ pkg/collector/corechecks/oracle/locks.go
