+++ pkg/collector/corechecks/oracle/os_stats.go
@@ -12,7 +12,7 @@ import (
 
 	"strings"
 
-	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/oracle-dbm/common"
+	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/oracle/common"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
