+++ pkg/collector/corechecks/oracle/processes.go
