+++ pkg/collector/corechecks/oracle/resource_manager.go
