+++ pkg/collector/corechecks/oracle/sender_util.go
