+++ pkg/collector/corechecks/oracle/shared_memory.go
