+++ pkg/collector/corechecks/oracle/sql_wrappers.go
