+++ pkg/collector/corechecks/oracle/statements.go
