+++ pkg/collector/corechecks/oracle/statements_test.go
