+++ pkg/collector/corechecks/oracle/stub.go
