+++ pkg/collector/corechecks/oracle/testutil.go
