+++ pkg/commonchecks/corechecks.go
@@ -29,7 +29,8 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/net/ntp"
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/networkpath"
 	nvidia "github.com/DataDog/datadog-agent/pkg/collector/corechecks/nvidia/jetson"
-	oracle "github.com/DataDog/datadog-agent/pkg/collector/corechecks/oracle-dbm"
+	oracle "github.com/DataDog/datadog-agent/pkg/collector/corechecks/oracle"
+	oracledbm "github.com/DataDog/datadog-agent/pkg/collector/corechecks/oracle-dbm"
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/orchestrator/pod"
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/sbom"
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/snmp"
@@ -78,6 +79,7 @@ func RegisterChecks(store workloadmeta.Component) {
 	corecheckLoader.RegisterCheck(network.CheckName, network.Factory())
 	corecheckLoader.RegisterCheck(nvidia.CheckName, nvidia.Factory())
 	corecheckLoader.RegisterCheck(oracle.CheckName, oracle.Factory())
+	corecheckLoader.RegisterCheck(oracledbm.CheckName, oracle.Factory())
 	corecheckLoader.RegisterCheck(disk.CheckName, disk.Factory())
 	corecheckLoader.RegisterCheck(wincrashdetect.CheckName, wincrashdetect.Factory())
 	corecheckLoader.RegisterCheck(winkmem.CheckName, winkmem.Factory())
