+++ tasks/build_tags.py
@@ -34,6 +34,7 @@ ALL_TAGS = {
     "netcgo",  # Force the use of the CGO resolver. This will also have the effect of making the binary non-static
     "npm",
     "oracle",
+    "oracle-dbm",
     "orchestrator",
     "otlp",
     "podman",
@@ -66,6 +67,7 @@ AGENT_TAGS = {
     "kubelet",
     "netcgo",
     "oracle",
+    "oracle-dbm",
     "orchestrator",
     "otlp",
     "podman",
@@ -88,6 +90,7 @@ AGENT_HEROKU_TAGS = AGENT_TAGS.difference(
         "kubeapiserver",
         "kubelet",
         "oracle",
+        "oracle-dbm",
         "orchestrator",
         "podman",
         "systemd",
